Search Result Page - Cambia nombre directorio y se renderiza card component
