+++ client/src/pages/search-result/styles.scss
@@ -1,0 +1,6 @@
+.page-search-result {
+  background-color: #fff;
+  padding: 16px;
+  border-radius: 4px;
+  margin-bottom: 120px;
+}
